time-funktion, switch-case
