--- Cosmos-Test-Platform/Kernel.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Sys = Cosmos.System;
-
-namespace Cosmos_Test_Platform
-{
-    public class Kernel : Sys.Kernel
-    {
-
-        protected override void BeforeRun()
-        {
-            Console.WriteLine("Cosmos booted successfully. Type a line of text to get it echoed back.");
-        }
-
-        protected override void Run()
-        {
-            Console.Write("Input: ");
-            var input = Console.ReadLine();
-            Console.Write("Text typed lolololollolol: ");
-            Console.WriteLine(input);
-        }
-    }
-}
